Move role package to controller-api
