+++ controller-api/src/main/java/com/yahoo/vespa/hosted/controller/api/role/Action.java
@@ -1,5 +1,5 @@
 // Copyright 2019 Oath Inc. Licensed under the terms of the Apache 2.0 license. See LICENSE in the project root.
-package com.yahoo.vespa.hosted.controller.role;
+package com.yahoo.vespa.hosted.controller.api.role;
 
 import com.yahoo.jdisc.http.HttpRequest;
