+++ controller-api/src/main/java/com/yahoo/vespa/hosted/controller/api/role/RolePrincipal.java
@@ -1,4 +1,4 @@
-package com.yahoo.vespa.hosted.controller.role;
+package com.yahoo.vespa.hosted.controller.api.role;
 
 import java.security.Principal;
